+++ sdk/storage/azure-storage-blob-nio/src/test/java/com/azure/storage/blob/nio/AttributeViewTest.groovy
@@ -6,6 +6,7 @@ package com.azure.storage.blob.nio
 import com.azure.storage.blob.BlobClient
 import com.azure.storage.blob.models.AccessTier
 import com.azure.storage.blob.models.BlobHttpHeaders
+import com.azure.storage.blob.models.BlobStorageException
 import spock.lang.Unroll
 
 import java.nio.file.ClosedFileSystemException
@@ -45,6 +46,7 @@ class AttributeViewTest extends APISpec {
         attr.isRegularFile()
         bc.getBlobUrl() == attr.fileKey()
         !attr.isDirectory()
+        !attr.isVirtualDirectory()
         !attr.isSymbolicLink()
         !attr.isOther()
     }
@@ -59,9 +61,39 @@ class AttributeViewTest extends APISpec {
 
         then:
         attr.isDirectory()
+        !attr.isVirtualDirectory()
         !attr.isRegularFile()
         !attr.isOther()
         !attr.isSymbolicLink()
+    }
+
+    def "AzureBasicFileAttributeView directory virtual"() {
+        setup:
+        def dirName = generateBlobName()
+        def childName = generateContainerName()
+        def bc = cc.getBlobClient(dirName + '/' + childName)
+        def path = fs.getPath(bc.getBlobName())
+
+        when:
+        def attr = new AzureBasicFileAttributeView(path).readAttributes()
+
+        then:
+        attr.isDirectory()
+        attr.isVirtualDirectory()
+        !attr.isRegularFile()
+        !attr.isOther()
+        !attr.isSymbolicLink()
+    }
+
+    def "AzureBasicFileAttributeView no exist"() {
+        setup:
+        def path = fs.getPath(generateBlobName())
+
+        when:
+        def attr = new AzureBasicFileAttributeView(path).readAttributes()
+
+        then:
+        thrown(IOException)
     }
 
     def "AzureBasicFileAttributeView fs closed"() {
@@ -93,6 +125,7 @@ class AttributeViewTest extends APISpec {
         attr.isRegularFile()
         bc.getBlobUrl() == attr.fileKey()
         !attr.isDirectory()
+        !attr.isVirtualDirectory()
         !attr.isSymbolicLink()
         !attr.isOther()
         props.getETag() == attr.eTag()
@@ -127,6 +160,7 @@ class AttributeViewTest extends APISpec {
         FileTime.from(props.getCreationTime().toInstant()) == suppliers.get("creationTime").get()
         attr.isRegularFile()
         !attr.isDirectory()
+        !attr.isVirtualDirectory()
         !attr.isSymbolicLink()
         !attr.isOther()
         props.getETag() == suppliers.get("eTag").get()